backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.1.3
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pyuploadcare==3.0.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
